feat(PSW-02-04): finish charts in a dashboard page
Finish the charts in the dashboard page, including the popular products
chart and the revenue chart. Also, add a 404 page for handling
non-existent routes.

Obs: The Card components had wrong background colors, and I fixed them
in a class base file and add Cartesian Grid in the charts for better
visualization and put messages to ignore verification about ts stuffs
on popular-chart component.