feat(PSW-03): create connection to show metrics and orders
Add API routes for restaurant management and user profile management.

Add positive interface because sometimes like a 98% of executions are
be successful and we want update that previous of the route can be
executed, but we need add something if the route fails.

- Implement pagination for order tables
- Create a dashboard layout for key order metrics
- Add API endpoints for fetching dashboard data

Create a datepicker to select a period you want to show in a chart. But
probably have bugs here. But I will correct this, because the class I
watched used a component like a DateRangePicker and it doenst exist
anymore.

Fix the values inside of chart because the values are in cents not in
reais.